Fission.CLI.Display.Error

Description

File sync, IPFS-style

Documentation

put :: (MonadIO m, MonadLogger m, Show err) => err -> Text -> m ()

Display a given error to the user and log an error to the debug log.

put' :: (MonadIO m, MonadLogger m, Show err) => err -> m ()

Display a generic error message to the user and log an error to the debug log.

putErrOr :: (MonadIO m, MonadLogger m, Show err) => (t -> m ()) -> Either err t -> m ()

notConnected :: (MonadIO m, MonadLogger m, Exception err) => err -> m ()

Display an error message to a user encouraging them to run `fission setup` Error depends on if they have basic auth saved somewhere (ie if they are an existing user)