Fission.CLI.Display.Loader

Description

Visual flourishes for indicating a loading state

Documentation

withLoader :: MonadUnliftIO m => Natural -> m a -> m a

Perform actions in the background while displaying a loading indicator

The indicator disappears when the process completes

reset :: MonadIO m => m ()

Reset the cursor position back one priontable character, and clear the *entire* line

prep :: MonadIO m => Natural -> m ()

Prepare for the next step -- in this case wait and reset the line

loading :: MonadIO m => Natural -> m ()

Loading animation