Fission.CLI.Environment

Description

Reading and writing local user config values

Documentation

init :: (MonadUnliftIO m, MonadLogger m, MonadWebClient m) => m ()

Initialize the Environment file

get :: MonadIO m => m Environment

Gets hierarchical environment by recursing through file system

getPath :: MonadIO m => Bool -> m FilePath

Get the path to the Environment file, local or global

couldNotRead :: MonadIO m => m ()

Create a could not read message for the terminal

removeConfigFile :: MonadUnliftIO m => m (Either IOException ())

Removes the user's global config file

getOrRetrievePeer :: (MonadUnliftIO m, MonadLogger m, MonadWebClient m) => Environment -> m (Maybe Peer)

Retrieves a Fission Peer from local config If not found we retrive from the network and store

Reexport

module Fission.CLI.Environment.Class

module Fission.CLI.Environment.Types