Fission.CLI.Environment.Override

Documentation

get :: MonadIO m => m Override

Gets hierarchical environment by recursed through file system

findBasicAuth :: MonadIO m => m (Maybe BasicAuthData)

Recurses up to user root to find an env with basic auth data

findRecurse :: MonadIO m => (Override -> Bool) -> FilePath -> m (Maybe (FilePath, Override))

Recurses up to user root to find a env that satisfies the finder predicate

decode :: MonadIO m => FilePath -> m Override

Decodes file to partial environment

write :: MonadIO m => FilePath -> Override -> m ()

Writes partial environment to path

writeMerge :: MonadIO m => FilePath -> Override -> m ()

Merge a partial env with an existing env at the path

globalEnv :: MonadIO m => m FilePath

globalEnv environment in users home

toFull :: Override -> Environment

fromFull :: Environment -> Override

deleteHomeAuth :: MonadIO m => m ()

Deletes user_auth from env at ~/.fission.yaml

module Fission.CLI.Environment.Override.Types