Fission.CLI.Environment.Partial

Documentation

get :: MonadIO m => m Partial

Gets hierarchical environment by recursed through file system

findBasicAuth :: MonadIO m => m (Maybe BasicAuthData)

Recurses up to user root to find an env with basic auth data

findRecurse :: MonadIO m => (Partial -> Bool) -> FilePath -> m (Maybe (FilePath, Partial))

Recurses up to user root to find a env that satisfies function "f"

decode :: MonadIO m => FilePath -> m Partial

Decodes file to partial environment

write :: MonadIO m => FilePath -> Partial -> m ()

Writes partial environment to path

writeMerge :: MonadIO m => FilePath -> Partial -> m ()

Merges partial env with the env at the path and overwrites

globalEnv :: MonadIO m => m FilePath

globalEnv environment in users home

toFull :: Partial -> Environment

fromFull :: Environment -> Partial

updatePeers :: Partial -> [Peer] -> Partial

deleteHomeAuth :: MonadIO m => m ()

Deletes user_auth from env at ~/.fission.yaml