Fission.CLI.Prompt

Documentation

reask :: (MonadIO m, MonadLogger m) => Text -> (ByteString -> Bool) -> m ByteString

Continues prompting the user until they put in a valid response

reaskNotEmpty :: (MonadIO m, MonadLogger m) => Text -> m ByteString

reask where valid responses are at least 1 character long

reaskNotEmpty' :: (MonadIO m, MonadLogger m) => Text -> m Text

reask where valid responses are at least 1 character long

reaskYN :: (MonadIO m, MonadLogger m) => Text -> m Bool

reask where valid responses are some form of yes/no

reaskWithError :: (MonadIO m, MonadLogger m) => Text -> (ByteString -> Bool) -> m () -> m ByteString