Fission.CLI.Prompt.Fields

Documentation

getRequired :: (MonadIO m, MonadLogger m) => ByteString -> m ByteString

Prompt a user for a value and do not accept an empty value

getRequiredSecret :: (MonadIO m, MonadLogger m) => ByteString -> m ByteString

Prompt a user for a secret and do not accept an empty value