Fission.Environment

Description

Environment variables

Documentation

withFlag :: String -> a -> a -> IO a

Switch on an environment flag

>>> withFlag "DEBUG" "nope" "yep"
"nope"

withEnv :: String -> a -> (String -> a) -> IO a

Perform actions on an environment variable, with fallback if not available

>>> withEnv "HOST" "my.host" (drop 1)
"my.host"

getFlag :: String -> IO (Maybe Bool)

Check if an environment flag is set to True (case-insensitive)

>>> getFlag "THIS_KEY_IS_UNSET"
Nothing

getEnv :: FromEnv a => IO a

Get an environment variable. errors if not found.

getFlagWithDefault :: String -> Bool -> IO Bool

Get a given flag, default to the given value if not found

(.!~) :: Monad m => m (Maybe a) -> a -> m a

Fallback value for a monadic lookup

>>> Right (Just 9) .!~ 42
Right 9

>>> Right Nothing .!~ 42
Right 42

>>> Left (Just 9) .!~ 42
Left (Just 9)

>>> Left Nothing .!~ 42
Left Nothing