Fission.Internal.UTF8

Description

UTF8 text helpers

Documentation

class Textable a where

Methods

encode :: a -> Either UnicodeException Text

putText :: MonadIO m => Text -> m ()

Helper for printing Text to a console

putTextLn :: MonadIO m => Text -> m ()

Helper for printing Text to a console with a newline at the end

displayLazyBS :: Display a => a -> ByteString

toBase58Text :: ByteString -> Text

Convert any binary object to Text

>>> toBase58Text "hello world"
"StV1DL6CwTryKyV"

>>> toBase58Text $ Strict.pack ([0x0ed, 0x01] :: [Word8] )
"K36"

>>> toBase58Text $ Strict.pack ([0xed, 0x01, 0x01, 0x23, 0x45, 0x67] :: [Word8])
"332DkaEge"

NOTE that base58 text does not concatenate without decoding to some base2 first

>>> toBase58Text "hello world" == toBase58Text "hello " <> toBase58Text "world"
False

fromRawBytes :: [Word8] -> Text

stripOptionalPrefix :: Text -> Text -> Text

stripOptionalPrefixBS :: ByteString -> ByteString -> ByteString

stripOptionalPrefixLazyBS :: ByteString -> ByteString -> ByteString

stripOptionalSuffix :: Text -> Text -> Text

stripOptionalSuffixBS :: ByteString -> ByteString -> ByteString

stripOptionalSuffixLazyBS :: ByteString -> ByteString -> ByteString

stripPadding :: ByteString -> ByteString

stripQuotes :: Text -> Text

stripQuotesBS :: ByteString -> ByteString

stripQuotesLazyBS :: ByteString -> ByteString

stripN :: Natural -> Text -> Text

Remove a number of characters from the beginning and the end of some text.

>>> stripN 3 "aaabccc"
"b"

>>> stripN 0 "b"
"b"

stripN n (Text.center (3 + fromIntegral n * 2) '_' "o.O") == "o.O"

stripNBS :: Natural -> ByteString -> ByteString

Remove a number of characters from the beginning and the end of a lazy ByteString.

>>> stripNBS 3 "aaabccc"
"b"

>>> stripNBS 0 "b"
"b"

stripNewline :: ByteString -> ByteString

Strip one newline character from the end of a lazy ByteString.

>>> stripNewline ";)\n"
";)"

>>> stripNewline "<>\n\n"
"<>\n"

stripNewline (Lazy.append bs "\n") == bs

stripNewline (Lazy.append bs "\n\n") == bs <> "\n"

textShow :: Show a => a -> Text

Show text.

>>> textShow 1
"1"

wrapIn :: Text -> Text -> Text

Wrap text with some other piece of text.

Text.head (wrapIn "|" s) == '|'

Text.last (wrapIn "|" s) == '|'

Text.length (wrapIn "|" s) == (Text.length s) + 2

wrapInBS :: ByteString -> ByteString -> ByteString