Fission.Platform.Heroku.Provision.Request.Types

Documentation

data Request

Constructors

Request

Fields

callbackUrl :: Text

The URL which should be used to retrieve updated information about the add-on and the app which owns it.

name :: Text

Logical name of the resource being provisioned. , oauthGrant :: Maybe Text -- OAuthGrant -- ^ OAuth object details (nullable).

plan :: Tier

Name of the plan to provision (e.g. basic).

region :: Region

Physical hosting region of the requesting client.

uuid :: UUID

The unique identifier Heroku uses for the installed add-on. It corresponds with the id field in the Heroku Platform API.