Fission.Platform.Heroku.Provision.Types

Documentation

data Provision

Response Parameters

From Heroku

In addition to the id parameter as documented under asynchronous provisioning,
you should include a config parameter along with any other optional parameters
of your choosing.

Name: config
Type: object

Description:
Configuration variables to be set as environment variables on any applications
that use this add-on resource. All environment variables that you send should
have the same prefix: your add-on’s name, capitalized. However,
within the context of a Heroku app, the prefix may be different for a variety
of reasons. For example, if your add-on is named fast-db and you are setting
FAST_DB_URL, the variable name on the application will default to FAST_DB_URL
but would be PRIMARY_DB_URL if the user added the add-on with
the prefix PRIMARY_DB.

Example:
HTTP/1.1 200 OK
{ "MYADDON_URL": "http://myaddon.com/52e82f5d73" }

Constructors

Provision

Fields

id :: UserId

User ID

config :: Provision

Heroku env var payload

peers :: [Peer]

IPFS peer list

message :: Text

A helpful human-readable message