Fission.Web.IPFS

Documentation

type API = AuthedAPI :<|> PublicAPI

type Auth = HigherOrder

type AuthedAPI = Auth :> UnauthedAPI

type PublicAPI = ("peers" :> API) :<|> API

type UnauthedAPI = ("cids" :> API) :<|> (API :<|> (API :<|> ("dag" :> API)))

authed :: (MonadRemoteIPFS m, MonadLocalIPFS m, MonadLogger m, MonadThrow m, MonadTime m, MonadDB t m, Creator t, Retriever t, Destroyer t) => ServerT AuthedAPI m

public :: (MonadLinkedIPFS m, MonadLocalIPFS m, MonadLogger m, MonadThrow m) => ServerT PublicAPI m

server :: (MonadRemoteIPFS m, MonadLinkedIPFS m, MonadLocalIPFS m, MonadLogger m, MonadThrow m, MonadTime m, MonadDB t m, Creator t, Retriever t, Destroyer t) => ServerT API m